refactor(cloister): extract specialists feedback seam (pan-2232)
